# storage/framework/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "Docs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base"
        }
    ],
    "paths": {
...
        "/t/{tenant}/comm/providers": {
            "get": {
                "tags": [
                    "Comm Providers"
                ],
                "summary": "List providers for this tenant",
                "operationId": "06b446c824d60702d34c63ea7cf60be3",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comm Providers"
                ],
                "summary": "Create provider (email|sms|push)",
                "operationId": "8a9a76bd864c4fbc90d53b23c59424d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "channel",
                                    "provider"
                                ],
                                "properties": {
                                    "channel": {
                                        "type": "string",
                                        "enum": [
                                            "email",
                                            "sms",
                                            "push"
                                        ]
                                    },
                                    "provider": {
                                        "type": "string",
                                        "example": "ses"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "credentials_json": {
                                        "type": "object"
                                    },
                                    "from_email": {
                                        "type": "string"
                                    },
                                    "from_name": {
                                        "type": "string"
                                    },
                                    "sender_id": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/t/{tenant}/comm/providers/{id}": {
            "put": {
                "tags": [
                    "Comm Providers"
                ],
                "summary": "Update provider",
                "operationId": "019b3d9084e310bcc40eb7c5a2f1039c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "credentials_json": {
                                        "type": "object"
                                    },
                                    "from_email": {
                                        "type": "string"
                                    },
                                    "from_name": {
                                        "type": "string"
                                    },
                                    "sender_id": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/t/{tenant}/roles": {
            "get": {
                "tags": [
                    "RBAC"
                ],
                "summary": "List roles in tenant",
                "operationId": "2af6292384c94955cc10d9ad2c16a7db",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "RBAC"
                ],
                "summary": "Create role in tenant",
                "operationId": "e77f47eca4ac46188e6c56e02aa9098e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/t/{tenant}/roles/{roleId}/permissions": {
            "post": {
                "tags": [
                    "RBAC"
                ],
                "summary": "Set permissions for a role (replace)",
                "operationId": "fa637b5cdfaaa1b4a326af82dad37868",
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permission_keys"
                                ],
                                "properties": {
                                    "permission_keys": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/t/{tenant}/users/{userId}/roles": {
            "post": {
                "tags": [
                    "RBAC"
                ],
                "summary": "Assign roles to user in tenant (replace)",
                "operationId": "7faea808468da7872fcfc2940efeb51e",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role_keys"
                                ],
                                "properties": {
                                    "role_keys": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/t/{tenant}/permissions": {
            "get": {
                "tags": [
                    "RBAC"
                ],
                "summary": "List available permission keys",
                "operationId": "59e3251fe008abb6924954a64cb72113",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}